+++ app/Http/Livewire/Series/UpcomingSeries.php
@@ -10,17 +10,18 @@ class UpcomingSeries extends Component
 {
     public $upcoming;
 
-    private string $search;
+    public string $search;
 
     protected $listeners = ['search' => 'filter'];
 
     public function render()
     {
-        $upcoming = Volume::with('series')->where('ignore_in_upcoming', 'false')->whereIn('status', [0, 1, 2])->orderBy('publish_date')->get();
+        $upcoming = Volume::with('series.publisher')->where('ignore_in_upcoming', 'false')->whereIn('status', [0, 1, 2])->orderBy('publish_date')->get();
         if (!empty($this->search)) {
             $upcoming = $upcoming->filter(function ($volume) {
                 return Str::contains(Str::lower($volume->name), Str::lower($this->search))
-                || Str::contains(Str::lower($volume->isbn), Str::lower($this->search));
+                || Str::contains(Str::lower($volume->isbn), Str::lower($this->search))
+                || Str::contains(Str::lower($volume->series->publisher?->name), Str::lower($this->search));
             });
         }
         $this->upcoming = $upcoming;
@@ -48,7 +49,7 @@ class UpcomingSeries extends Component
         $volume = Volume::find($id);
         $volume->status = $status;
         $volume->save();
-        $this->emitTo('global-statistics', '$refresh');
+        $this->emitTo('overview', '$refresh');
         toastr()->livewire()->addSuccess(__(':name has been updated', ['name' => $volume->series->name . ' ' . $volume->number]));
     }