+++ resources/views/livewire/series/gallery.blade.php
@@ -1,6 +1,6 @@
 <div class="bg-white shadow-sm rounded my-2 px-3 py-2">
     <div class="my-2">
-        <h1 style="display: inline;">{{ $category->name }}</h1>
+        <h1 style="display: inline;">{{ $category->name }} <small class="text-muted">({{ count($series) }} {{ __('SeriesPlural') }})</small></h1>
         <div class="float-end">
             @if ($category->type == 0)
                 <a href="{{ route('series.create', [$category]) }}" class="btn btn-link"><i class="fas fa-plus-circle"></i></a>
@@ -17,7 +17,7 @@
             </div>
         @else
             @foreach ($series as $item)
-                @include('series.card', [$item])
+                @include('series.card', [$category, $item])
             @endforeach
         @endif
     </div>